[Update] 1/N Unified the registration of custom operators to torch.ops and fixed some minor issues
